Move init_logging into pimm.logging, so an entry point and a spawned child configure alike
`positronic.utils.logging` was the entry point's half of a configuration whose other half is
`pimm.logging`, and it had already been reduced to a colour call around it. Split across two
modules the two halves drift by construction — that is how only children came to pin the noisy
libraries — so `init_logging` moves in beside `configure_process_logging`, and the file goes.
`coloredlogs` comes with it into `pimm`: a hard dependency of the distribution `pimm` ships in,
and the only thing an entry point does that a child does not, since a child's output is a pipe.

25 import sites repointed, mechanically. `positronic/utils/tests/test_logging.py` merges into
`pimm/tests/test_logging.py`, which now covers both calls in one module with one fixture — the
cross-file leak the last commit fixed cannot be written again between files that no longer
exist apart.

Ticket: Positronic-Robotics/internal#561 #open
