+++ positronic/cli/__init__.py
@@ -1,9 +1,9 @@
 import configuronic as cfn
 import pos3
 
+from pimm.logging import init_logging
 from positronic.cli.account import commands as account_commands
 from positronic.cli.eval import commands as eval_commands
-from positronic.utils.logging import init_logging
 
 
 @pos3.with_mirror()
