+++ positronic/vendors/lerobot/server.py
@@ -6,6 +6,7 @@ from typing import Any
 import configuronic as cfn
 import pos3
 
+from pimm.logging import init_logging
 from positronic import keys
 from positronic.offboard.server import serve
 from positronic.offboard.server_utils import run_with_progress, warmup
@@ -14,7 +15,6 @@ from positronic.policy.codec import RestrictImageSize
 from positronic.policy.spec import ModelSource, Pipeline, remote
 from positronic.policy.wrappers import ChunkedSchedule, StopOnFault
 from positronic.utils.checkpoints import list_checkpoints, resolve_checkpoint
-from positronic.utils.logging import init_logging
 from positronic.vendors.lerobot import codecs as lerobot_codecs
 from positronic.vendors.lerobot.policy import LerobotPolicy, _detect_device, warm_observation
