+++ positronic/vendors/lerobot/to_lerobot.py
@@ -24,10 +24,10 @@ import torch
 import tqdm
 from lerobot.datasets.lerobot_dataset import LeRobotDataset
 
+from pimm.logging import init_logging
 from positronic import utils
 from positronic.cfg.ds import apply_codec
 from positronic.dataset import Dataset
-from positronic.utils.logging import init_logging
 
 
 def _raise_fd_limit(min_soft_limit: int = 4096) -> None:
